+++ packages/yume-engine/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './staleness-tracker.js';
+export * from './timeout.js';